Support multi-mode lexing for preprocessors
