+++ packages/language/src/language-server/completion/keywords.ts
@@ -191,7 +191,8 @@ export const CompletionKeywords = {
     [CstNodeKind.DetachStatement_DETACH, kw("DETACH")],
     [CstNodeKind.DisplayStatement_DISPLAY, kw("DISPLAY")],
     [CstNodeKind.DoStatement_DO, kw("DO")],
-    [CstNodeKind.ExecStatement_EXEC, kw("EXEC")],
+    [CstNodeKind.ExecCicsStatement_EXEC, kw("EXEC")],
+    [CstNodeKind.ExecSqlStatement_EXEC, kw("EXEC")],
     [CstNodeKind.ExitStatement_EXIT, kw("EXIT")],
     [CstNodeKind.FetchStatement_FETCH, kw("FETCH")],
     [CstNodeKind.FlushStatement_FLUSH, kw("FLUSH")],
