+++ packages/language/src/syntax-tree/cst.ts
@@ -270,9 +270,12 @@ export enum CstNodeKind {
   DoSpecification_UPTHRU,
   DoSpecification_DOWNTHRU,
   DoSpecification_REPEAT,
-  ExecStatement_EXEC,
-  ExecStatement_Query,
+  ExecStatement_Token,
   ExecStatement_Semicolon,
+  ExecCicsStatement_EXEC,
+  ExecCicsStatement_CICS,
+  ExecSqlStatement_EXEC,
+  ExecSqlStatement_SQL,
   ExitStatement_EXIT,
   ExitStatement_Semicolon,
   FetchStatement_FETCH,
