+++ packages/language/test/fourslash/preprocessor/cics/exec-generates-declarations.ts
@@ -68,7 +68,5 @@ TEST: PROC;
     DCL DFHEI01 ENTRY OPTIONS(INTER ASSEMBLER);
 
     DCL VAR1 FIXED BIN(31);
-    EXEC CICS SOMETHING1;
-    EXEC CICS SOMETHING2;
 END;
 `);
