+++ packages/vscode-extension/syntaxes/pli.manual.json
@@ -36,15 +36,6 @@
     "string-character-escape": {
       "name": "constant.character.escape.pli",
       "match": "\\\\(x[0-9A-Fa-f]{2}|u[0-9A-Fa-f]{4}|u\\{[0-9A-Fa-f]+\\}|[0-2][0-7]{0,2}|3[0-6][0-7]?|37[0-7]?|[4-7][0-7]?|.|$)"
-    },
-    "comments": {
-      "patterns": [
-        {
-          "name": "string.exec.pli",
-          "begin": "(?i)(?<=EXEC\\s+)[a-z]+\\s",
-          "end": "(?=;)"
-        }
-      ]
     }
   }
 }
